Написана своя реализация связанного списка
